finished moving around some message related pieces
